Fix #77 - The issue with the method call of `removeFile()` without passing the file has been fixed

+++ projects/iplab/ngx-file-upload/src/lib/helpers/control.class.ts
@@ -102,6 +102,10 @@ export class FileUploadControl {
     }
 
     public removeFile(file: File): this {
+        if (IsNullOrEmpty(file)) {
+            throw Error(`File has not been provided.`);
+        }
+
         if (!this.disabled) {
             this.files.delete(file.name);
             this.validate();
